[Jenkins] Set User, Mail for Jenkins in meta App
Previously we had default values in Jenkins Application directly, but
these defaults are used only for demo and testing. Current patch remove
these default values and move them to meta CICD Application.

Currently it's the same emal and user as used for authorization in
Jenkins and Gerrit. In the future it may be separate user/mail for
independent work Jenkins with Gerrit.

Change-Id: Id13c407225f4fbe766567207a40fb606bb04a085